Move UiWindow into a separate file
