+++ src/viewer/viewer.h
@@ -31,7 +31,7 @@
 
 #include <memory>
 
-#include "viewer/ui/window.h"
+#include "viewer/ui/ui_window.h"
 
 namespace viewer {
